refactor base react app setup
