--- packages/client/src/env-config.ts
@@ -1,2 +1,0 @@
-//@ts-ignore
-;(globalThis as any).process = { env: { ...(import.meta as any).env, APP_ENV: (import.meta as any).env.MODE } }
