+++ packages/client/src/main.tsx
@@ -1,62 +1,19 @@
-// prettier-ignore
-import './env-config';
-
-import React, { Suspense, useEffect } from 'react'
+import React, { Suspense } from 'react'
 import ReactDOM from 'react-dom'
 
-import { API } from '@xrengine/client-core/src/API'
-import { FullscreenContainer } from '@xrengine/client-core/src/components/FullscreenContainer'
 import { LoadingCircle } from '@xrengine/client-core/src/components/LoadingCircle'
-import { createEngine, initializeBrowser, setupEngineActionSystems } from '@xrengine/engine/src/initializeEngine'
 
-import { initializei18n } from './util'
+//@ts-ignore
+;(globalThis as any).process = { env: { ...(import.meta as any).env, APP_ENV: (import.meta as any).env.MODE } }
 
-/**
- * Performance benchmark logging
- */
+const Engine = React.lazy(() => import('./engine'))
 
-// const log = console.log
-// console.log = (...args) => log(`${Math.round(performance.now()/100) / 10}s`, ...args)
-
-// const info = console.info
-// console.info = (...args) => info(`${Math.round(performance.now()/100) / 10}s`, ...args)
-
-// const warn = console.warn
-// console.warn = (...args) => warn(`${Math.round(performance.now()/100) / 10}s`, ...args)
-
-// const error = console.error
-// console.error = (...args) => error(`${Math.round(performance.now()/100) / 10}s`, ...args)
-
-const AppPage = React.lazy(() => import('./pages/_app'))
-
-const canvasStyle = {
-  zIndex: -1,
-  width: '100%',
-  height: '100%',
-  position: 'fixed',
-  WebkitUserSelect: 'none',
-  pointerEvents: 'auto',
-  userSelect: 'none'
-} as React.CSSProperties
-const engineRendererCanvasId = 'engine-renderer-canvas'
-
-const Main = () => {
-  useEffect(() => {
-    createEngine()
-    setupEngineActionSystems()
-    initializeBrowser()
-    API.createAPI()
-  }, [])
-
+const App = () => {
   return (
     <Suspense fallback={<LoadingCircle />}>
-      <FullscreenContainer>
-        <canvas id={engineRendererCanvasId} style={canvasStyle} />
-        <AppPage />
-      </FullscreenContainer>
+      <Engine />
     </Suspense>
   )
 }
 
-initializei18n()
-ReactDOM.render(<Main />, document.getElementById('root'))
+ReactDOM.render(<App />, document.getElementById('root'))
